+++ circle.go
@@ -35,9 +35,9 @@ func (c Circle[T]) Resize(radius T) Circle[T] {
 	return Circle[T]{c.Center, radius}
 }
 
-// Grow creates a new Circle with radius increased by amount.
+// Grow creates a new Circle with radius increased by amount, clamped to zero.
 func (c Circle[T]) Grow(amount T) Circle[T] {
-	return Circle[T]{c.Center, c.Radius + amount}
+	return Circle[T]{c.Center, max(c.Radius+amount, 0)}
 }
 
 // Shrink creates a new Circle with radius decreased by amount, clamped to zero.