+++ types/ints/types.go
@@ -61,7 +61,7 @@ func Pad[T geom.Number](top, right, bottom, left T) Padding {
 	return geom.Pad(top, right, bottom, left).Int()
 }
 
-// Mat is shorthand for geom.Matrix(a, b, c, d, e, f).Int()
+// Mat is shorthand for geom.Mat(a, b, c, d, e, f).Int()
 func Mat[T geom.Number](a, b, c, d, e, f T) Matrix {
 	return geom.Mat(a, b, c, d, e, f).Int()
 }